♻️ Refactor: 코드 분리 및 Tanstack Query 도입
